Fix debug assertion failure on Visual Studio
When compiling and running texrecon in debug mode on Visual Studio, it exits
due to a failed debug assertion. Specifically, "vector subscript out of range".
This is because BVHTree::ssplit() in libs/acc/bvh_tree.h passes a pointer as the
second argument to std::sort() which can point one element past the end of the
vector like an iterator would. Using iterators instead of pointers fixes the
failed assertion.